pypi structure
