+++ additional-scripts/plot_effect_of_gamma.py
