+++ pmoss/__init__.py
@@ -1,0 +1,2 @@
+from .version import __version__
+from .utils import group_names, group_combination, create_combination
