+++ pmoss/analysis/__init__.py
@@ -1,0 +1,1 @@
+from .data_diagnosis import get_decision_index, data_diagnosis, compute_diagnosis
