+++ pmoss/display.py
@@ -10,10 +10,39 @@ import pandas as pd
 import matplotlib.pyplot as plt
 import matplotlib as mpl
 import seaborn as sns
-from exponential_fit import decission_data_exponential
-from lowess_fit import significance_analysis
+from .models.exponential_fit import decission_data_exponential
+from .models.lowess_fit import significance_analysis
+from .analysis import get_decision_index
 
-def func_exp_pure(x, a, c):    
+def table_of_results(param, data_features, combination_dict):
+    Theta = get_decision_index(param, data_features, combination_dict)
+
+    for i in range(len(data_features)):
+        t = Theta[['comparison', data_features[np.str(i)] + ' Theta']]
+        p = param[['comparison', data_features[np.str(i)] + '_nalpha_estimated',
+                   data_features[np.str(i)] + '_nalpha_theory']]
+        p = p.assign(a=np.nan, c=np.nan)
+        p[['a', 'c']] = param[data_features[np.str(i)] + '_exp_params'].apply(
+            lambda x: pd.Series([x[0], x[1]], index=['a', 'c']))
+        p = p[['comparison', 'a', 'c',
+               data_features[np.str(i)] + '_nalpha_estimated',
+               data_features[np.str(i)] + '_nalpha_theory']]
+        p.rename(columns={'a': data_features[np.str(i)] + ' a',
+                          'c': data_features[np.str(i)] + ' c',
+                          data_features[np.str(i)] + '_nalpha_estimated':
+                              data_features[np.str(i)] + ' ^n-alpha',
+                          data_features[np.str(i)] + '_nalpha_theory':
+                              data_features[np.str(i)] + ' n-alpha'},
+                 inplace=True)
+
+        if i == 0:
+            table = pd.merge(p, t, on='comparison')
+        else:
+            aux = pd.merge(p, t, on='comparison')
+            table = pd.merge(table, aux, on='comparison')
+    return table
+
+def func_exp_pure(x, a, c):
     return a*np.exp(-c*(x))
               
 def plot_decission_with_LOWESS(df, combination,test, measure, fs = None,
