+++ pmoss/statistics/__init__.py
@@ -1,0 +1,2 @@
+from .cross_validation import cross_validated_pvalues
+from .statistical_tests import compare_continuous_samples, pvalues_continuous, obtain_pvalues_ChiSquared
